Separate groovy and java because sigh
