+++ gradle-baseline-java/src/main/groovy/com/palantir/baseline/plugins/XmlUtils.java
@@ -14,27 +14,24 @@
  * limitations under the License.
  */
 
-package com.palantir.baseline.plugins
+package com.palantir.baseline.plugins;
 
-import com.google.common.collect.ImmutableMap
-import com.google.common.io.Files
-import java.nio.charset.Charset
-import java.util.function.Consumer
-import javax.xml.parsers.ParserConfigurationException
-import org.xml.sax.SAXException
+import com.google.common.collect.ImmutableMap;
+import com.google.common.io.Files;
+import groovy.util.Node;
+import groovy.util.XmlNodePrinter;
+import groovy.util.XmlParser;
+import java.io.BufferedWriter;
+import java.io.File;
+import java.io.IOException;
+import java.io.PrintWriter;
+import java.nio.charset.StandardCharsets;
+import java.util.function.Consumer;
+import javax.xml.parsers.ParserConfigurationException;
+import org.xml.sax.SAXException;
 
 final class XmlUtils {
-    static Node matchOrCreateChild(Node base, String name, Map attributes = [:], Map defaults = [:],
-                                   @DelegatesTo(Node) Closure ifCreated = {}) {
-        def child = base[name].find { it.attributes().entrySet().containsAll(attributes.entrySet()) }
-        if (child) {
-            return child
-        }
-
-        def created = base.appendNode(name, attributes + defaults)
-        ifCreated(created)
-        return created
-    }
+    private XmlUtils() {}
 
     static void createOrUpdateXmlFile(File configurationFile, Consumer<Node> configure) {
         Node rootNode;
@@ -48,10 +45,10 @@ final class XmlUtils {
             rootNode = new Node(null, "project", ImmutableMap.of("version", "4"));
         }
 
-        configure(rootNode);
+        configure.accept(rootNode);
 
-        try (BufferedWriter writer = Files.newWriter(configurationFile, Charset.defaultCharset());
-             PrintWriter printWriter = new PrintWriter(writer)) {
+        try (BufferedWriter writer = Files.newWriter(configurationFile, StandardCharsets.UTF_8);
+                PrintWriter printWriter = new PrintWriter(writer)) {
             XmlNodePrinter nodePrinter = new XmlNodePrinter(printWriter);
             nodePrinter.setPreserveWhitespace(true);
             nodePrinter.print(rootNode);
